backend/requirements.txt
django>=4.2,<4.3
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.2,<6.0
django-cors-headers>=4.3,<5.0